refactored code
conditional operators replaced if-else's

renamed 'showSideBar' to 'toggleSidebar'

split NavBar into smaller components

removed unused styled component (Checkbox)
